requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
deep-translator==1.11.4
deepl==1.18.0
filelock==3.16.0
fsspec==2024.9.0
fuzzywuzzy==0.18.0
google-trans-new==1.1.9
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
joblib==1.4.2
langdetect==1.0.9
Levenshtein==0.25.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
packaging==24.1
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-telegram-bot==21.5
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
rfc3986==1.5.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.3
pandas==2.0.1
seaborn==0.12.2
matplotlib==3.7.1
